+++ initsystem/initsystem.go
@@ -25,19 +25,68 @@ import (
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
+const (
+	_STATUS_UNKNOWN     = 0
+	_STATUS_PRESENT     = 1
+	_STATUS_NOT_PRESENT = 2
+)
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+var (
+	sysvStatus    = _STATUS_UNKNOWN
+	upstartStatus = _STATUS_UNKNOWN
+	systemdStatus = _STATUS_UNKNOWN
+)
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
 // SysV returns true if SysV is used on system
 func SysV() bool {
-	return !Systemd()
+	if sysvStatus != _STATUS_UNKNOWN {
+		return sysvStatus == _STATUS_PRESENT
+	}
+
+	switch Systemd() {
+	case true:
+		sysvStatus = _STATUS_NOT_PRESENT
+	default:
+		sysvStatus = _STATUS_PRESENT
+	}
+
+	return sysvStatus == _STATUS_PRESENT
 }
 
 // Upstart returns true if Upstart is used on system
 func Upstart() bool {
-	return env.Which("initctl") != ""
+	if upstartStatus != _STATUS_UNKNOWN {
+		return upstartStatus == _STATUS_PRESENT
+	}
+
+	switch env.Which("initctl") {
+	case "":
+		upstartStatus = _STATUS_NOT_PRESENT
+	default:
+		upstartStatus = _STATUS_PRESENT
+	}
+
+	return upstartStatus == _STATUS_PRESENT
 }
 
 // Systemd returns true if Systemd is used on system
 func Systemd() bool {
-	return env.Which("systemctl") != ""
+	if systemdStatus != _STATUS_UNKNOWN {
+		return systemdStatus == _STATUS_PRESENT
+	}
+
+	switch env.Which("systemctl") {
+	case "":
+		systemdStatus = _STATUS_NOT_PRESENT
+	default:
+		systemdStatus = _STATUS_PRESENT
+	}
+
+	return systemdStatus == _STATUS_PRESENT
 }
 
 // IsPresent returns true if service is present in any init system